requirements-dev.txt
pytest==7.4.0
pytest-watch==4.2.0
ruff==0.1.8
pre-commit==3.5.0
python-dotenv==1.0.0 